requirements.txt
# Core dependencies
torch==2.9.0
transformers @ git+https://github.com/huggingface/transformers
accelerate>=0.33.0
tokenizers>=0.15.0
safetensors>=0.4.3

# Dataset handling
datasets>=2.0.0
pandas==1.4.0

# Utilities
numpy==1.22.1
tqdm==4.62.3
PyYAML==6.0
pillow>=10.0.0
ipywidgets>=8.0.0  # Required for tqdm progress bars in Jupyter notebooks

# Editor dependencies
higher==0.2.1  # Used by MEND editor

# Optional - comment out if not using ROME editor
# matplotlib==3.5.1  # Used by ROME compute_v